fix branch instructions

+++ src/witness/operation.rs
@@ -35,7 +35,7 @@ pub(crate) enum BranchCond {
 }
 
 impl BranchCond {
-    pub(crate) fn result(&self, input0: usize, input1: usize) -> bool {
+    pub(crate) fn result(&self, input0: i32, input1: i32) -> bool {
         match self {
             BranchCond::EQ => input0 == input1,
             BranchCond::NE => input0 != input1,
@@ -390,7 +390,7 @@ pub(crate) fn generate_branch<F: Field>(
 ) -> Result<(), ProgramError> {
     let (src1, src1_op) = reg_read_with_log(src1, 0, state, &mut row)?;
     let (src2, src2_op) = reg_read_with_log(src2, 1, state, &mut row)?;
-    let should_jump = cond.result(src1, src2);
+    let should_jump = cond.result(src1 as i32, src2 as i32);
     reg_write_with_log(0, 2, src1 - src2, state, &mut row)?;
     reg_write_with_log(0, 3, src2 - src1, state, &mut row)?;
     let pc = state.registers.program_counter as u32;
